machine: add accton-csp7551.conf
The Accton CSP7551 is an x86 platform with a Tofino chip. The CPU can
exchange data with the Tofino through two dedicated, internal e810
Ethernet controllers.

There are multiple models of the CSP7551 with different CPU models and
RAM and storage configurations.

The CSP7551 has 32 external 100 GB ports, a console port, and a RJ-45
port used as a shared management port by the host and the BMC (IPMI).

Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>