csp7551: add platform init script
The platform init script loads the drivers and performs the
i2c configuration.

The file name is "platform-$(onie-sysinfo -p | tr '_' '-')-init.sh".

The i2c configuration is taken from [1]; instead of removing the
i2c-i801 module from the kernel, we assume it is not loaded.

[1] https://github.com/ShoneWu/sonic-buildimage/blob/202205_fd5ddba_csp7551/platform/barefoot/sonic-platform-modules-accton/csp7551/utils/accto
n_csp7551_util.py

Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>